scenarios/identity-theft/yamlidator/requirements.txt
flask
flask_restful
pyyaml
cerberus